requirements.txt
# =========================================================
# ClassWatch — Production Requirements
# Stable AI / CV deployment environment
# Optimized for Flask + SocketIO threading mode
# =========================================================

# ---------------------------------------------------------
# Deep Learning / GPU
# ---------------------------------------------------------
# Install CUDA builds separately if needed:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121

# ---------------------------------------------------------
# Computer Vision / AI Pipeline
# ---------------------------------------------------------
ultralytics==8.4.37
ultralytics-thop==2.0.18

mediapipe==0.10.14

opencv-python==4.13.0.92
opencv-contrib-python==4.13.0.92

# ---------------------------------------------------------
# Numerical / Scientific Computing
# ---------------------------------------------------------
numpy==2.4.4
scipy==1.17.1

jax==0.9.2
jaxlib==0.9.2

# ---------------------------------------------------------
# Web Dashboard / Backend
# ---------------------------------------------------------
Flask==3.1.3
Flask-SocketIO==5.6.1

Werkzeug==3.1.8

python-engineio==4.13.1
python-socketio==5.16.1

simple-websocket==1.1.0
wsproto==1.3.2

# Production server
gunicorn==23.0.0

# Environment configuration
python-dotenv==1.1.1

# ---------------------------------------------------------
# Analytics / Visualization
# ---------------------------------------------------------
matplotlib==3.10.8

contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
kiwisolver==1.5.0
pyparsing==3.3.2

# ---------------------------------------------------------
# Utilities
# ---------------------------------------------------------
requests==2.33.1

python-dateutil==2.9.0.post0
PyYAML==6.0.3

psutil==7.2.2
packaging==26.0

typing_extensions==4.15.0

filelock==3.25.2
fsspec==2026.3.0

sympy==1.13.1
networkx==3.6.1

pillow==12.2.0

# ---------------------------------------------------------
# Data Processing
# ---------------------------------------------------------
polars==1.39.3
polars-runtime-32==1.39.3

# ---------------------------------------------------------
# HTTP / Networking
# ---------------------------------------------------------
urllib3==2.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11

# ---------------------------------------------------------
# Flask Internals
# ---------------------------------------------------------
blinker==1.9.0
click==8.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3

# ---------------------------------------------------------
# Misc Dependencies
# ---------------------------------------------------------
absl-py==2.4.0
attrs==26.1.0
bidict==0.23.1

cffi==2.0.0
pycparser==3.0

colorama==0.4.6

flatbuffers==25.12.19

h11==0.16.0

lap==0.5.13

ml_dtypes==0.5.4

mpmath==1.3.0

opt_einsum==3.4.0

protobuf==4.25.9

setuptools==81.0.0
six==1.17.0

